+++ src/vmm/src/devices/virtio/iovec.rs
@@ -324,7 +324,8 @@ mod tests {
     use super::{IoVecBuffer, IoVecBufferMut};
     use crate::devices::virtio::queue::{Queue, VIRTQ_DESC_F_NEXT, VIRTQ_DESC_F_WRITE};
     use crate::devices::virtio::test_utils::VirtQueue;
-    use crate::vstate::memory::{Bytes, GuestAddress, GuestMemoryExtension, GuestMemoryMmap};
+    use crate::utilities::test_utils::multi_region_mem;
+    use crate::vstate::memory::{Bytes, GuestAddress, GuestMemoryMmap};
 
     impl<'a> From<&'a [u8]> for IoVecBuffer {
         fn from(buf: &'a [u8]) -> Self {
@@ -371,15 +372,11 @@ mod tests {
     }
 
     fn default_mem() -> GuestMemoryMmap {
-        GuestMemoryMmap::from_raw_regions(
-            &[
-                (GuestAddress(0), 0x10000),
-                (GuestAddress(0x20000), 0x10000),
-                (GuestAddress(0x40000), 0x10000),
-            ],
-            false,
-        )
-        .unwrap()
+        multi_region_mem(&[
+            (GuestAddress(0), 0x10000),
+            (GuestAddress(0x20000), 0x10000),
+            (GuestAddress(0x40000), 0x10000),
+        ])
     }
 
     fn chain(m: &GuestMemoryMmap, is_write_only: bool) -> (Queue, VirtQueue) {