+++ src/vmm/src/devices/virtio/mmio.rs
@@ -364,7 +364,8 @@ pub(crate) mod tests {
 
     use super::*;
     use crate::devices::virtio::ActivateError;
-    use crate::vstate::memory::{GuestMemoryExtension, GuestMemoryMmap};
+    use crate::utilities::test_utils::single_region_mem;
+    use crate::vstate::memory::GuestMemoryMmap;
 
     #[derive(Debug)]
     pub(crate) struct DummyDevice {
@@ -465,7 +466,7 @@ pub(crate) mod tests {
 
     #[test]
     fn test_new() {
-        let m = GuestMemoryMmap::from_raw_regions(&[(GuestAddress(0), 0x1000)], false).unwrap();
+        let m = single_region_mem(0x1000);
         let mut dummy = DummyDevice::new();
         // Validate reset is no-op.
         assert!(dummy.reset().is_none());
@@ -497,7 +498,7 @@ pub(crate) mod tests {
 
     #[test]
     fn test_bus_device_read() {
-        let m = GuestMemoryMmap::from_raw_regions(&[(GuestAddress(0), 0x1000)], false).unwrap();
+        let m = single_region_mem(0x1000);
         let mut d = MmioTransport::new(m, Arc::new(Mutex::new(DummyDevice::new())), false);
 
         let mut buf = vec![0xff, 0, 0xfe, 0];
@@ -586,7 +587,7 @@ pub(crate) mod tests {
     #[test]
     #[allow(clippy::cognitive_complexity)]
     fn test_bus_device_write() {
-        let m = GuestMemoryMmap::from_raw_regions(&[(GuestAddress(0), 0x1000)], false).unwrap();
+        let m = single_region_mem(0x1000);
         let dummy_dev = Arc::new(Mutex::new(DummyDevice::new()));
         let mut d = MmioTransport::new(m, dummy_dev.clone(), false);
         let mut buf = vec![0; 5];
@@ -745,7 +746,7 @@ pub(crate) mod tests {
 
     #[test]
     fn test_bus_device_activate() {
-        let m = GuestMemoryMmap::from_raw_regions(&[(GuestAddress(0), 0x1000)], false).unwrap();
+        let m = single_region_mem(0x1000);
         let mut d = MmioTransport::new(m, Arc::new(Mutex::new(DummyDevice::new())), false);
 
         assert!(!d.are_queues_valid());
@@ -863,7 +864,7 @@ pub(crate) mod tests {
 
     #[test]
     fn test_bus_device_reset() {
-        let m = GuestMemoryMmap::from_raw_regions(&[(GuestAddress(0), 0x1000)], false).unwrap();
+        let m = single_region_mem(0x1000);
         let mut d = MmioTransport::new(m, Arc::new(Mutex::new(DummyDevice::new())), false);
         let mut buf = [0; 4];