+++ src/vmm/src/devices/virtio/vsock/event_handler.rs
@@ -225,7 +225,8 @@ mod tests {
     use super::*;
     use crate::devices::virtio::vsock::packet::VSOCK_PKT_HDR_SIZE;
     use crate::devices::virtio::vsock::test_utils::{EventHandlerContext, TestContext};
-    use crate::vstate::memory::{Bytes, GuestMemoryExtension, GuestMemoryMmap};
+    use crate::utilities::test_utils::multi_region_mem;
+    use crate::vstate::memory::Bytes;
 
     #[test]
     fn test_txq_event() {
@@ -475,15 +476,11 @@ mod tests {
         const MIB: usize = 1 << 20;
 
         let mut test_ctx = TestContext::new();
-        test_ctx.mem = GuestMemoryMmap::from_raw_regions(
-            &[
-                (GuestAddress(0), 8 * MIB),
-                (GuestAddress((GAP_START_ADDR - MIB) as u64), MIB),
-                (GuestAddress(FIRST_AFTER_GAP as u64), MIB),
-            ],
-            false,
-        )
-        .unwrap();
+        test_ctx.mem = multi_region_mem(&[
+            (GuestAddress(0), 8 * MIB),
+            (GuestAddress((GAP_START_ADDR - MIB) as u64), MIB),
+            (GuestAddress(FIRST_AFTER_GAP as u64), MIB),
+        ]);
 
         // The default configured descriptor chains are valid.
         {